Updated article and comment layout with styling and structure improvements
